requirements.txt
Django>=3.2,<4.0
mysqlclient>=2.0
djangorestframework
django-cors-headers
djangorestframework-simplejwt
